requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
Pillow==8.0.1
pkg-resources==0.0.0
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
